requirements.txt
environ-config==22.1.*
influxdb-client==1.35.*
pymindergas==0.1.*